refactor: 액티비티 클래스 ui 패키지로 이동
