+++ app/src/main/java/com/moroom/android/ui/ResultActivity.java
@@ -1,4 +1,4 @@
-package com.moroom.android;
+package com.moroom.android.ui;
 
 
 import android.content.Intent;
@@ -14,6 +14,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.moroom.android.MoveToLogin;
+import com.moroom.android.R;
 import com.moroom.android.data.model.Review;
 import com.moroom.android.databinding.ActivityResultBinding;
 import com.google.firebase.auth.FirebaseAuth;
