+++ app/src/main/java/com/moroom/android/ui/navui/home/HomeFragment.kt
@@ -17,9 +17,9 @@ import com.google.firebase.auth.FirebaseAuth
 
 import com.moroom.android.MoveToLogin
 import com.moroom.android.R
-import com.moroom.android.ResultActivity
-import com.moroom.android.SearchActivity
-import com.moroom.android.WriteActivity
+import com.moroom.android.ui.ResultActivity
+import com.moroom.android.ui.SearchActivity
+import com.moroom.android.ui.WriteActivity
 import com.moroom.android.databinding.FragmentHomeBinding
 
 class HomeFragment : Fragment() {
