+++ app/src/main/res/layout/activity_write.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="10dp"
-    tools:context=".WriteActivity">
+    tools:context=".ui.WriteActivity">
 
     <TextView
         android:layout_width="wrap_content"
